refactor(asp/forms): BirthPlaceAndCountryMixin
defines commonly used fields for jobseeker creation and hides the complexity of managing birth place, country and birthdate with INSEE records

refactor(forms): further simplification
